content/html/content/test/test_bug421640.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=421640
-->
<head>
  <title>Test for Bug 421640</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=421640">Mozilla Bug 421640</a>
<div id="content">
  <div id="edit" contenteditable="true">This text is editable</div>
  <div><button id="button">Test</button></div> 
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 421640 **/

function test(click, focus, nextFocus) {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  var wu =  window.QueryInterface(Components.interfaces.nsIInterfaceRequestor)
                  .getInterface(Components.interfaces.nsIDOMWindowUtils);

  var selection = window.getSelection();
  var edit = document.getElementById("edit");
  var text = edit.firstChild;

  selection.removeAllRanges();

  var rect = edit.getBoundingClientRect();
  wu.sendMouseEvent("mousedown", rect.left + 1, rect.top + 1, 0, 1, 0);
  wu.sendMouseEvent("mousemove", rect.right - 1, rect.top + 1, 0, 1, 0);
  wu.sendMouseEvent("mouseup", rect.right - 1, rect.top + 1, 0, 1, 0);

  is(selection.anchorNode, text, "");

  rect = document.getElementById("button").getBoundingClientRect();
  wu.sendMouseEvent("mousedown", rect.left + 10, rect.top + 1, 0, 1, 0);
  wu.sendMouseEvent("mouseup", rect.left + 10, rect.top + 1, 0, 1, 0);

  is(selection.anchorNode, text, "");

  SimpleTest.finish();
}

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  setTimeout(test, 0);
};

</script>
</pre>
</body>
</html>
